+++ lib/core.dart
@@ -20,6 +20,7 @@ library;
 export 'src/components/seo_component_format.dart';
 export 'src/components/seo_components.dart';
 export 'src/components/seo_collection_transition.dart';
+export 'src/components/seo_collection_url.dart';
 export 'src/components/seo_motion.dart' show SeoMotionPreset;
 export 'src/components/seo_rich_text.dart';
 export 'src/components/seo_tabs_transition.dart';